Game Engine that is really in an alpha phase as there are many bugs and a good amount of features do not work completely well.
The program does use all cross-platform libraries so it is possible to
compile natively on Linux and Mac OSX but for now the engine is easily
available/compileable for Windows only.

An image shows the needed files and structure to make sure nothing is
missing. Python 2.7 is also needed in which it is looking for default
location which on windows is usually C:\Python27.

Should compile easily with visual studio 2013 (12) and has been test compiled
with visual studio 2013 (12)